expose exec apis
